src/main/java/com/hhplus/ecommerce/business/ItemService.java
package com.hhplus.ecommerce.business;

import com.hhplus.ecommerce.domain.Item;
import com.hhplus.ecommerce.infrastructure.ItemRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@AllArgsConstructor
public class ItemService {

    private final ItemRepository itemRepository;

    public Item save(Item item) {
        return itemRepository.save(item);
    }

    public List<Item> getAllItems() {

        return itemRepository.findAll();
    }

    public Item findById(long itemId) {

        return itemRepository.findById(itemId).orElse(null);
    }
}
